+++ src/main/java/org/realityforge/replicant/client/json/gwt/ReplicantConfig.java
@@ -8,4 +8,6 @@ public interface ReplicantConfig
   extends PropertySource
 {
   boolean shouldValidateRepositoryOnLoad();
+
+  boolean repositoryDebugOutputEnabled();
 }